hive_api fixes, reorg
